email-monitor/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
redis==5.2.1
pydantic==2.10.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
openai>=1.0.0
